Split project team component to a standalone file
